temp

+++ mesh/inc/meshx_lower_transport.h
@@ -12,9 +12,10 @@
 
 MESHX_BEGIN_DECLS
 
+
 MESHX_EXTERN int32_t meshx_lower_transport_init(void);
 MESHX_EXTERN int32_t meshx_lower_transport_send(meshx_network_if_t network_if,
-                                                const uint8_t *pupper_trans_pdu, uint8_t len, meshx_msg_ctx_t *pmsg_ctx);
+                                                const uint8_t *pupper_trans_pdu, uint16_t len, meshx_msg_ctx_t *pmsg_ctx);
 
 MESHX_END_DECLS
